text-summarization-app/requirements.txt
flask
nltk
gensim
transformers
beautifulsoup4
requests
scikit-learn
pandas
numpy
PyPDF2
rouge-score
bleu-score